requirements.txt
alembic==1.6.5
attrs==21.2.0
bleach==3.2.1
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
cfscrape==2.1.1
chardet==4.0.0
click==7.1.2
colorthief==0.2.1
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.3.2
enum34==1.1.10
feedparser==5.2.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
idna==2.10
importlib-metadata==2.1.1
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==1.1.1
packaging==20.9
pathlib2==2.3.6
pika==1.2.0
Pillow==6.2.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==3.12
raven==6.10.0
requests==2.26.0
scandir==1.10.0
sender==0.3
service-identity==16.0.0
six==1.16.0
SQLAlchemy==1.4.23
subprocess32==3.5.4
Twisted==16.0.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==1.0.1
zipp==1.2.0
zope.interface==5.4.0
